backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-environ==0.11.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
Pillow==11.1.0
whitenoise==6.8.2
